requirements.txt
flask>=3.0
waitress>=2.1
gunicorn>=21.0
werkzeug>=3.0
flask-cors>=4.0
flask-session>=0.5
flask-sqlalchemy>=3.1
sqlalchemy>=2.0
alembic>=1.13
authlib>=1.3
pyjwt>=2.8
flask-limiter>=3.5
email-validator>=2.1
argon2-cffi>=23.1
google-generativeai>=0.3
gTTS>=2.5
langdetect>=1.0.9
requests>=2.31
beautifulsoup4>=4.12
python-dotenv>=1.0
Pillow>=10.2
redis
black>=24.0
psycopg2-binary>=2.9
celery>=5.3
gevent>=23.9
python-magic>=0.4.27
